+++ agentloop-20260520/src/Models/UpdateExperimentPlanRequest.php
@@ -46,6 +46,11 @@ class UpdateExperimentPlanRequest extends Model
     /**
      * @var string
      */
+    public $pipelineName;
+
+    /**
+     * @var string
+     */
     public $planName;
 
     /**
@@ -65,6 +70,7 @@ class UpdateExperimentPlanRequest extends Model
         'experimentType' => 'experimentType',
         'experiments' => 'experiments',
         'input' => 'input',
+        'pipelineName' => 'pipelineName',
         'planName' => 'planName',
         'querySql' => 'querySql',
         'selectedItemIds' => 'selectedItemIds',
@@ -135,6 +141,10 @@ class UpdateExperimentPlanRequest extends Model
                     $res['input'][$key1] = $value1;
                 }
             }
+        }
+
+        if (null !== $this->pipelineName) {
+            $res['pipelineName'] = $this->pipelineName;
         }
 
         if (null !== $this->planName) {
@@ -212,6 +222,10 @@ class UpdateExperimentPlanRequest extends Model
                     $model->input[$key1] = $value1;
                 }
             }
+        }
+
+        if (isset($map['pipelineName'])) {
+            $model->pipelineName = $map['pipelineName'];
         }
 
         if (isset($map['planName'])) {