+++ agentloop-20260520/ChangeLog.md
@@ -1,3 +1,19 @@
+2026-07-30 Version: 2.1.0
+- Support API CreateExperimentPlan.
+- Support API CreateExperimentRun.
+- Support API DeleteExperimentPlan.
+- Support API DeleteExperimentRun.
+- Support API GetExperimentPlan.
+- Support API GetExperimentRun.
+- Support API ListExperimentPlans.
+- Support API ListExperimentRuns.
+- Support API UpdateExperimentPlan.
+- Support API UpdateExperimentRun.
+- Update API GetDataset: add response parameters Body.labels.
+- Update API ListDatasets: add request parameters labels.
+- Update API ListDatasets: add response parameters Body.datasets.$.labels.
+
+
 2026-07-23 Version: 2.1.0
 - Support API CreateExperimentPlan.
 - Support API CreateExperimentRun.